+++ src/main/java/net/staticstudios/prisons/pickaxe/enchants/DoubleJackHammerEnchant.java
@@ -3,11 +3,10 @@ package net.staticstudios.prisons.pickaxe.enchants;
 import net.staticstudios.prisons.pickaxe.enchants.handler.BaseEnchant;
 import net.staticstudios.prisons.pickaxe.enchants.handler.EnchantTier;
 
-import java.math.BigInteger;
 
 public class DoubleJackHammerEnchant extends BaseEnchant {
     public DoubleJackHammerEnchant() {
-        super("doubleJackhammer", "&6&lDouble Wammy", 1000, BigInteger.valueOf(7500), "&7Chance to destroy an additional mine layer", "Activates along side Jack Hammer");
+        super("doubleJackhammer", "&6&lDouble Wammy", 1000, 7500, "&7Chance to destroy an additional mine layer", "Activates along side Jack Hammer");
         setPickaxeLevelRequirement(70);
 
         setTiers(
